Combine #[cgp_handler] into #[cgp_computer]
