+++ crates/cgp-extra/src/prelude.rs
@@ -1,4 +1,4 @@
-pub use cgp_core::macros::{cgp_computer, cgp_handler, cgp_producer};
+pub use cgp_core::macros::{cgp_computer, cgp_producer};
 pub use cgp_handler::{
     AsyncComputer, AsyncComputerComponent, AsyncComputerRef, AsyncComputerRefComponent, Computer,
     ComputerComponent, ComputerRefComponent, Handler, HandlerComponent, HandlerRefComponent,
