+++ crates/cgp-macro-lib/src/entrypoints/mod.rs
@@ -4,7 +4,6 @@ mod cgp_component;
 mod cgp_computer;
 mod cgp_context;
 mod cgp_getter;
-mod cgp_handler;
 mod cgp_new_provider;
 mod cgp_preset;
 mod cgp_producer;
@@ -26,7 +25,6 @@ pub use cgp_component::*;
 pub use cgp_computer::*;
 pub use cgp_context::*;
 pub use cgp_getter::*;
-pub use cgp_handler::*;
 pub use cgp_new_provider::*;
 pub use cgp_preset::*;
 pub use cgp_producer::*;
