+++ crates/cgp-macro/src/lib.rs
@@ -951,13 +951,6 @@ pub fn cgp_computer(attr: TokenStream, body: TokenStream) -> TokenStream {
         .into()
 }
 
-#[proc_macro_attribute]
-pub fn cgp_handler(attr: TokenStream, body: TokenStream) -> TokenStream {
-    cgp_macro_lib::cgp_handler(attr.into(), body.into())
-        .unwrap_or_else(syn::Error::into_compile_error)
-        .into()
-}
-
 #[proc_macro_derive(HasField)]
 pub fn derive_fields(item: TokenStream) -> TokenStream {
     cgp_macro_lib::derive_fields(item.into()).into()
